+++ Modules/Microsoft365DSC/DscResources/MSFT_IntuneDeviceConfigurationCustomPolicyMacOS/MSFT_IntuneDeviceConfigurationCustomPolicyMacOS.psm1
@@ -1,0 +1,425 @@
+using module ..\_Base\M365DSCResourceBase.psm1
+
+[DscResource()]
+class IntuneDeviceConfigurationCustomPolicyMacOS : M365DSCResourceBase
+{
+    [DscProperty()]
+    [System.ComponentModel.Description('Represents the assignment to the Intune policy.')]
+    [MSFT_DeviceManagementConfigurationPolicyAssignments[]] $Assignments
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Indicates the channel used to deploy the configuration profile. Available choices are DeviceChannel, UserChannel. Possible values are: deviceChannel, userChannel, unknownFutureValue.')]
+    [ValidateSet('deviceChannel', 'userChannel', 'unknownFutureValue')]
+    [System.String] $DeploymentChannel
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Admin provided description of the Device Configuration.')]
+    [System.String] $Description
+
+    [DscProperty(Key)]
+    [System.ComponentModel.Description('Admin provided name of the device configuration.')]
+    [System.String] $DisplayName
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The unique identifier for an entity. Read-only.')]
+    [System.String] $Id
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Payload. (UTF8 encoded byte array)')]
+    [System.String] $Payload
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Payload file name (.mobileconfig')]
+    [System.String] $PayloadFileName
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Name that is displayed to the user.')]
+    [System.String] $PayloadName
+
+    [DscProperty()]
+    [System.ComponentModel.Description('List of Scope Tags for this Entity instance.')]
+    [System.String[]] $RoleScopeTagIds
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Specify if the policy should exist.')]
+    [ValidateSet('Present', 'Absent')]
+    [System.String] $Ensure
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Credentials of the Intune Admin')]
+    [System.Management.Automation.PSCredential] $Credential
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Id of the Entra ID application to authenticate with.')]
+    [System.String] $ApplicationId
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Id of the Entra ID tenant used for authentication.')]
+    [System.String] $TenantId
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Secret of the Entra ID application to authenticate with.')]
+    [System.Management.Automation.PSCredential] $ApplicationSecret
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Thumbprint of the Entra ID application''s authentication certificate to use for authentication.')]
+    [System.String] $CertificateThumbprint
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Username can be made up to anything but password will be used for CertificatePassword')]
+    [System.Management.Automation.PSCredential] $CertificatePassword
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Path to certificate used in service principal usually a PFX file.')]
+    [System.String] $CertificatePath
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Managed ID being used for authentication.')]
+    [System.Nullable[System.Boolean]] $ManagedIdentity
+
+    [DscProperty()]
+    [System.ComponentModel.Description('Access token used for authentication.')]
+    [System.String[]] $AccessTokens
+
+    # Export-only. Not part of the resource schema.
+    [System.String] $Filter
+
+    [IntuneDeviceConfigurationCustomPolicyMacOS] Get()
+    {
+        if ($this.RequiresPowerShellCore())
+        {
+            $remote = [IntuneDeviceConfigurationCustomPolicyMacOS]::new()
+            $remote.FromHashtable($this.InvokeInPowerShellCore('Get'))
+            return $remote
+        }
+
+        Write-Verbose -Message "Getting configuration of Intune Device Configuration Custom Policy for macOS {$($this.Id)}"
+
+        try
+        {
+            $null = $this.Connect('MicrosoftGraph')
+
+            #Ensure the proper dependencies are installed in the current environment.
+            Confirm-M365DSCDependencies
+
+            #region Telemetry
+            $this.AddTelemetry('Get')
+            #endregion
+
+            $nullResult = $this.GetBoundParameters()
+            $nullResult.Ensure = 'Absent'
+
+            if (-not $this.ExportedInstance -or $this.ExportedInstance.Id -ne $this.Id)
+            {
+                $getValue = $null
+                if (-not [System.String]::IsNullOrEmpty($this.Id))
+                {
+                    $getValue = Get-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $this.Id `
+                        -ErrorAction SilentlyContinue
+                }
+
+                if ($null -eq $getValue -and -not [System.String]::IsNullOrEmpty($this.DisplayName))
+                {
+                    $getValue = Get-MgBetaDeviceManagementDeviceConfiguration `
+                        -Filter "DisplayName eq '$($this.DisplayName -replace "'", "''")' and isof('microsoft.graph.macOSCustomConfiguration')" `
+                        -ErrorAction SilentlyContinue | Select-Object -First 1
+                }
+            }
+            else
+            {
+                $getValue = $this.ExportedInstance
+            }
+
+            if ($null -eq $getValue)
+            {
+                Write-Verbose -Message "No Intune Device Configuration Custom Policy for macOS with Id {$($this.Id)} was found"
+                return $this.AsResult($nullResult)
+            }
+
+            Write-Verbose -Message "Found Intune Device Configuration Custom Policy for macOS with Id {$($this.Id)}"
+
+            $enumDeploymentChannel = $null
+            if ($null -ne $getValue.deploymentChannel)
+            {
+                $enumDeploymentChannel = $getValue.deploymentChannel.ToString()
+            }
+            $result = @{
+                DeploymentChannel     = $enumDeploymentChannel
+                Description           = $getValue.Description
+                DisplayName           = $getValue.DisplayName
+                Id                    = $getValue.Id
+                Payload               = $this.DecodeTextPayload($getValue.payload)
+                PayloadFileName       = $getValue.payloadFileName
+                PayloadName           = $getValue.payloadName
+                RoleScopeTagIds       = Resolve-M365DSCIntuneRoleScopeTagNames -CurrentValues $getValue.RoleScopeTagIds -DesiredValues $this.RoleScopeTagIds
+                Ensure                = 'Present'
+                Credential            = $this.Credential
+                ApplicationId         = $this.ApplicationId
+                TenantId              = $this.TenantId
+                ApplicationSecret     = $this.ApplicationSecret
+                CertificateThumbprint = $this.CertificateThumbprint
+                CertificatePassword   = $this.CertificatePassword
+                CertificatePath       = $this.CertificatePath
+                ManagedIdentity       = $this.ManagedIdentity
+                AccessTokens          = $this.AccessTokens
+            }
+
+            $assignmentsValues = Get-M365DSCIntuneExpandedAssignments -Instance $getValue
+            if ($null -eq $assignmentsValues)
+            {
+                $assignmentsValues = Get-MgBetaDeviceManagementDeviceConfigurationAssignment -DeviceConfigurationId $getValue.Id -ErrorAction SilentlyContinue
+            }
+            $assignmentResult = @()
+            if ($null -ne $assignmentsValues -and $assignmentsValues.Count -gt 0)
+            {
+                $assignmentResult += ConvertFrom-IntunePolicyAssignment `
+                    -IncludeDeviceFilter:$true `
+                    -Assignments $assignmentsValues
+            }
+            $result.Add('Assignments', $assignmentResult)
+
+            return $this.AsResult($result)
+        }
+        catch
+        {
+            $this.LogError($_, 'Error retrieving data:')
+
+            throw
+        }
+    }
+
+    [void] Set()
+    {
+        if ($this.RequiresPowerShellCore())
+        {
+            $null = $this.InvokeInPowerShellCore('Set')
+            return
+        }
+
+        Write-Verbose -Message "Setting configuration of Intune Device Configuration Custom Policy for macOS {$($this.Id)}"
+
+        Confirm-M365DSCDependencies
+
+        $this.AddTelemetry('Set')
+
+        try
+        {
+            $null = $this.Connect('MicrosoftGraph')
+
+            $currentInstance = $this.Get().ToHashtable()
+
+            $boundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $this.GetBoundParameters()
+
+            if ($boundParameters.ContainsKey('Payload'))
+            {
+                $boundParameters.Payload = $this.EncodeTextPayload($this.Payload)
+            }
+
+            if ($boundParameters.ContainsKey('RoleScopeTagIds'))
+            {
+                $boundParameters.RoleScopeTagIds = Resolve-M365DSCIntuneRoleScopeTagIds -RoleScopeTagIds $this.RoleScopeTagIds
+            }
+
+            $boundParameters = Rename-M365DSCCimInstanceParameter -Properties $boundParameters
+            $boundParameters.Remove('Id') | Out-Null
+            $boundParameters.Remove('Assignments') | Out-Null
+            $boundParameters.Add('@odata.type', '#microsoft.graph.macOSCustomConfiguration')
+
+            if ($this.Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
+            {
+                Write-Verbose -Message "Creating new Intune Device Configuration Custom Policy for macOS {$($this.Id)}"
+
+                $createParameters = $boundParameters
+                $createdInstance = New-MgBetaDeviceManagementDeviceConfiguration -BodyParameter $createParameters
+
+                $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $this.Assignments
+                if ($createdInstance.Id)
+                {
+                    Update-DeviceConfigurationPolicyAssignment `
+                        -DeviceConfigurationPolicyId $createdInstance.Id `
+                        -Targets $assignmentsHash `
+                        -Repository 'deviceManagement/deviceConfigurations'
+                }
+            }
+            elseif ($this.Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
+            {
+                Write-Verbose -Message "Updating Intune Device Configuration Custom Policy for macOS {$($this.Id)}"
+
+                $updateParameters = $boundParameters
+                Update-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $currentInstance.Id -BodyParameter $updateParameters | Out-Null
+
+                $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $this.Assignments
+                if ($currentInstance.Id)
+                {
+                    Update-DeviceConfigurationPolicyAssignment `
+                        -DeviceConfigurationPolicyId $currentInstance.Id `
+                        -Targets $assignmentsHash `
+                        -Repository 'deviceManagement/deviceConfigurations'
+                }
+            }
+            elseif ($this.Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
+            {
+                Write-Verbose -Message "Removing Intune Device Configuration Custom Policy for macOS {$($this.Id)}"
+                Remove-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $currentInstance.Id | Out-Null
+            }
+        }
+        catch
+        {
+            $this.LogError($_, 'Error updating data:')
+
+            throw
+        }
+    }
+
+    [bool] Test()
+    {
+        return ([M365DSCResourceBase] $this).Test()
+    }
+
+    [string] Export()
+    {
+        if ($this.RequiresPowerShellCore())
+        {
+            return [string] $this.InvokeInPowerShellCore('Export')
+        }
+
+        $ConnectionMode = $this.Connect('MicrosoftGraph')
+
+        Confirm-M365DSCDependencies
+
+        $this.AddTelemetry('Export')
+
+        try
+        {
+            [array] $exportedInstances = Get-M365DSCExportCachedCollection -Collection 'deviceConfigurations' `
+                -ODataType 'microsoft.graph.macOSCustomConfiguration' `
+                -Filter $this.Filter
+
+            $dscContent = [System.Text.StringBuilder]::new()
+            $i = 1
+
+            if ($exportedInstances.Length -eq 0)
+            {
+                Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark -CommitWrite
+            }
+            else
+            {
+                Write-M365DSCHost -Message "`r`n" -DeferWrite
+            }
+
+            foreach ($exportedInstance in $exportedInstances)
+            {
+                if ($null -ne $Global:M365DSCExportResourceInstancesCount)
+                {
+                    $Global:M365DSCExportResourceInstancesCount++
+                }
+
+                Write-M365DSCHost -Message "    |---[$i/$($exportedInstances.Count)] $($exportedInstance.DisplayName)" -DeferWrite
+
+                $Params = @{
+                    Id                    = $exportedInstance.Id
+                    DisplayName           = $exportedInstance.DisplayName
+                    Credential            = $this.Credential
+                    ApplicationId         = $this.ApplicationId
+                    TenantId              = $this.TenantId
+                    ApplicationSecret     = $this.ApplicationSecret
+                    CertificateThumbprint = $this.CertificateThumbprint
+                    CertificatePassword   = $this.CertificatePassword
+                    CertificatePath       = $this.CertificatePath
+                    ManagedIdentity       = $this.ManagedIdentity
+                    AccessTokens          = $this.AccessTokens
+                }
+
+                $this.ExportedInstance = $exportedInstance
+                $Results = $this.GetForExport($Params)
+
+                if ($null -ne $Results.Assignments)
+                {
+                    $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString `
+                        -ComplexObject $Results.Assignments `
+                        -CIMInstanceName 'MSFT_DeviceManagementConfigurationPolicyAssignments'
+                    if (-not [System.String]::IsNullOrWhiteSpace($complexTypeStringResult))
+                    {
+                        $Results.Assignments = $complexTypeStringResult
+                    }
+                    else
+                    {
+                        $Results.Remove('Assignments') | Out-Null
+                    }
+                }
+                $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $this.GetResourceName() `
+                    -ConnectionMode $ConnectionMode `
+                    -ModulePath $this.GetModulePath() `
+                    -Results $Results `
+                    -Credential $this.Credential `
+                    -NoEscape @('Assignments')
+                [void]$dscContent.Append($currentDSCBlock)
+                Save-M365DSCPartialExport -Content $currentDSCBlock `
+                    -FileName $Global:PartialExportFileName
+
+                Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark -CommitWrite
+                $i++
+            }
+
+            return $dscContent.ToString()
+        }
+        catch
+        {
+            Write-M365DSCHost -Message $Global:M365DSCEmojiRedX
+
+            $this.LogError($_, 'Error during Export:')
+
+            throw
+        }
+    }
+
+    hidden [IntuneDeviceConfigurationCustomPolicyMacOS] AsResult([System.Object] $Values)
+    {
+        if ($Values -is [IntuneDeviceConfigurationCustomPolicyMacOS])
+        {
+            return $Values
+        }
+
+        $result = [IntuneDeviceConfigurationCustomPolicyMacOS]::new()
+        $result.ClearNonSchemaProperties()
+        if ($Values -is [System.Collections.Hashtable])
+        {
+            $result.FromHashtable($Values)
+        }
+
+        return $result
+    }
+}
+
+class MSFT_DeviceManagementConfigurationPolicyAssignments
+{
+    [DscProperty(Mandatory)]
+    [System.ComponentModel.Description('The type of the target assignment.')]
+    [ValidateSet('#microsoft.graph.cloudPcManagementGroupAssignmentTarget', '#microsoft.graph.groupAssignmentTarget', '#microsoft.graph.allLicensedUsersAssignmentTarget', '#microsoft.graph.allDevicesAssignmentTarget', '#microsoft.graph.exclusionGroupAssignmentTarget', '#microsoft.graph.configurationManagerCollectionAssignmentTarget')]
+    [System.String] $dataType
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The type of filter of the target assignment i.e. Exclude or Include. Possible values are:none, include, exclude.')]
+    [ValidateSet('none', 'include', 'exclude')]
+    [System.String] $deviceAndAppManagementAssignmentFilterType
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The Id of the filter for the target assignment.')]
+    [System.String] $deviceAndAppManagementAssignmentFilterId
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The display name of the filter for the target assignment.')]
+    [System.String] $deviceAndAppManagementAssignmentFilterDisplayName
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The group Id that is the target of the assignment.')]
+    [System.String] $groupId
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The group Display Name that is the target of the assignment.')]
+    [System.String] $groupDisplayName
+
+    [DscProperty()]
+    [System.ComponentModel.Description('The collection Id that is the target of the assignment.(ConfigMgr)')]
+    [System.String] $collectionId
+}